+++ packages/wallets/adapters/ton/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @layerswap/wallet-ton
 
+## 2.2.0
+
+### Minor Changes
+
+- Global improvements and bug fixes
+
+### Patch Changes
+
+- Updated dependencies
+  - @layerswap/utils@2.2.0
+  - @layerswap/wallet-core@2.2.0
+  - @layerswap/widget-types@2.2.0
+
 ## 2.1.0
 
 ### Minor Changes